talking-jellyfish/models/object-detection/requirements.txt
seldon-core
transformers
timm
Pillow